+++ tests/test_convert_libre_office.py
@@ -12,14 +12,13 @@ from gotenberg_client.options import PdfAFormat
 from tests.conftest import SAMPLE_DIR
 from tests.conftest import SAVE_DIR
 from tests.conftest import SAVE_OUTPUTS
-from tests.utils import call_run_with_server_error_handling
 
 
 class TestLibreOfficeConvert:
     def test_libre_office_convert_docx_format(self, client: GotenbergClient):
         test_file = SAMPLE_DIR / "sample.docx"
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(route.convert(test_file))
+            resp = route.convert(test_file).run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers
@@ -31,7 +30,7 @@ class TestLibreOfficeConvert:
     def test_libre_office_convert_odt_format(self, client: GotenbergClient):
         test_file = SAMPLE_DIR / "sample.odt"
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(route.convert(test_file))
+            resp = route.convert(test_file).run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers
@@ -43,7 +42,7 @@ class TestLibreOfficeConvert:
     def test_libre_office_convert_xlsx_format(self, client: GotenbergClient):
         test_file = SAMPLE_DIR / "sample.xlsx"
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(route.convert(test_file))
+            resp = route.convert(test_file).run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers
@@ -55,7 +54,7 @@ class TestLibreOfficeConvert:
     def test_libre_office_convert_ods_format(self, client: GotenbergClient):
         test_file = SAMPLE_DIR / "sample.ods"
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(route.convert(test_file))
+            resp = route.convert(test_file).run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers
@@ -66,8 +65,8 @@ class TestLibreOfficeConvert:
 
     def test_libre_office_convert_multiples_format(self, client: GotenbergClient):
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(
-                route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"]).no_merge(),
+            resp = (
+                route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"]).no_merge().run_with_retry()
             )
 
         assert resp.status_code == codes.OK
@@ -79,9 +78,7 @@ class TestLibreOfficeConvert:
 
     def test_libre_office_convert_multiples_format_merged(self, client: GotenbergClient):
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(
-                route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"]).merge(),
-            )
+            resp = route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"]).merge().run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers
@@ -94,8 +91,10 @@ class TestLibreOfficeConvert:
         with patch("gotenberg_client._utils.guess_mime_type") as mocked_guess_mime_type:
             mocked_guess_mime_type.side_effect = guess_mime_type_stdlib
             with client.libre_office.to_pdf() as route:
-                resp = call_run_with_server_error_handling(
-                    route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"]).no_merge(),
+                resp = (
+                    route.convert_files([SAMPLE_DIR / "sample.docx", SAMPLE_DIR / "sample.odt"])
+                    .no_merge()
+                    .run_with_retry()
                 )
 
             assert resp.status_code == codes.OK
@@ -117,7 +116,7 @@ class TestLibreOfficeConvert:
     ):
         test_file = SAMPLE_DIR / "sample.xlsx"
         with client.libre_office.to_pdf() as route:
-            resp = call_run_with_server_error_handling(route.convert(test_file).pdf_format(gt_format))
+            resp = route.convert(test_file).pdf_format(gt_format).run_with_retry()
 
         assert resp.status_code == codes.OK
         assert "Content-Type" in resp.headers